// back/src/datagateway/model/student/IStateModel.ts
interface IStateModel {
  salary: boolean;
  plainTime: boolean;
  middleTime: boolean;
  CDD: boolean;
  CDI: boolean;
  INTERIM: boolean;
  help: boolean;
  employmentAsker: boolean;
  homeChildren: boolean;
  countryAsker: boolean;
  home: boolean;
  AAH: boolean;
  ESAT: boolean;
  youngAlone: boolean;
  other: boolean;
  ESATDetails: string;
  employmentAskerDate: Date;
  otherDetails: string;
  comment: string;
  RSA: boolean;
}

export default IStateModel;